requirements.txt
nulltype==2.3.1
numpy==2.2.1
pandas==2.2.3
plaid-python==28.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
tzdata==2024.2
urllib3==2.3.0
